% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gen_fr.R
\name{simulate_gen_fr}
\alias{simulate_gen_fr}
\title{simulate_gen_fr: simulates a generalized functional response time series}
\usage{
simulate_gen_fr(
  n_initial,
  p = 1,
  f_max,
  n_half,
  q,
  t_start = 0,
  t_end = 1,
  t_length = 1000
)
}
\arguments{
\item{n_initial}{integer or float; a vector of initial prey densities.}

\item{p}{integer or float; a single value of a fixed predator density. The
default value is 1.}

\item{f_max}{float; maximum feeding rate, a single value.}

\item{n_half}{float; half saturation density, a single value.}

\item{q}{float; shape parameter, a single value. A strict type II functional
has q = 0, a strict type III functional response has q = 1.}

\item{t_start}{integer or float; the time were the feeding starts. A single
value; default = 0.}

\item{t_end}{integer or float; the time were the feeding ends. A single
value; default = 1 (e.g. 1 day).}

\item{t_length}{integer or float; the number of time steps that should be
generated. The more time steps, the more precise the simulation. A single
value; default = 1000.}
}
\value{
Returns a data frame with n_initial and the according n_eaten.
}
\description{
\code{simulate_gen_fr} simulates time series across different initial
resource densities using \code{odin} (FitzJohn and Jombart 2024). It returns
the number of initial prey items and the number of prey eaten at the end
of the time series, mimicking common functional response laboratory
experiments. The underlying functional response model is the generalized
functional response model (Real 1977, 1979). Note that not integers, but
floating numbers are returned by \code{simulate_gen_fr}.
}
\examples{

# simulate the functional response:
out <- simulate_gen_fr(
 n_initial = 1:100,    # vector of initial prey densities
 p = 1,                # fixed predator density
 f_max = 10,           # maximum feeding rate
 n_half = 25,          # half saturation density
 q = 1                 # shape parameter (1 = s-shaped)
)

# plot the functional response
plot(
  out$n_initial,
  out$n_eaten,
  type = "l"
)

}
\references{
FitzJohn and Jombart (2024) odin: ODE generation and integration.
Ver. 1.2.6. https://doi.org/10.32614/CRAN.package.odin
see also: https://github.com/mrc-ide/odin

Real (1977) The kinetics of functional response. Am Nat 111, 289-
300. https://doi.org/10.1086/283161

Real (1979) Ecological determinants of functional response.
Ecology 60, 481-485. https://doi.org/10.2307/1936067
}
\seealso{
\code{\link[=compile_gen_fr]{compile_gen_fr()}}
}
